Improved interface and logging.
There is now an automatic port selection function, so that we do not
rely on a possibly free hardcoded port.
The logger now allows to catch all messages happening on the server, and
has a parameter to control loglevel. It also uses separate loggers for
each file, rather than the root logger for everyone.